fix: resolve merge conflicts from service-based architecture merge
- Add missing GitCliApi methods (fetchRemotes, listBranches, listWorktrees,
  getCurrentBranch, addExistingBranchWorktree, createTrackingBranchWorktree,
  branchExistsLocally, revParse, isAncestor, listRemotes) that were in the
  broken git_cli_api_commands.dart extension (which accessed private _processRunner)
- Remove git_cli_api_commands.dart (was an extension on GitCliApi that couldn't
  access private members, so none of its methods ever worked)
- Add isDedicated to WorktreeSuccess constructor and wire through services
- Export BranchInfo, BranchListResponse, WorktreeMode from sesori_shared
- Fix toSharedSession() to accept optional hasWorktree instead of branchName
- Fix session_archive_status_service to not pass branchName to toSharedSession
- Fix all WorktreeSuccess instantiations to include isDedicated parameter
