+++ bridge/app/lib/src/bridge/api/git_cli_api.dart
@@ -233,6 +233,93 @@ class GitCliApi {
     return _processRunner.run("git", arguments, workingDirectory: projectPath);
   }
 
+  Future<ProcessResult> fetchRemotes({required String workingDirectory}) {
+    return _processRunner.run("git", ["fetch", "--all"], workingDirectory: workingDirectory);
+  }
+
+  Future<ProcessResult> listBranches({required String workingDirectory}) {
+    return _processRunner.run(
+      "git",
+      ["branch", "-a", "--format=%(refname:short) %(creatordate:unix)"],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> listWorktrees({required String workingDirectory}) {
+    return _processRunner.run("git", ["worktree", "list", "--porcelain"], workingDirectory: workingDirectory);
+  }
+
+  Future<ProcessResult> getCurrentBranch({required String workingDirectory}) {
+    return _processRunner.run("git", ["rev-parse", "--abbrev-ref", "HEAD"], workingDirectory: workingDirectory);
+  }
+
+  Future<ProcessResult> addExistingBranchWorktree({
+    required String workingDirectory,
+    required String worktreePath,
+    required String branchName,
+  }) {
+    return _processRunner.run(
+      "git",
+      ["worktree", "add", "--", worktreePath, branchName],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> createTrackingBranchWorktree({
+    required String workingDirectory,
+    required String worktreePath,
+    required String localBranchName,
+    required String remoteBranch,
+  }) {
+    return _processRunner.run(
+      "git",
+      ["worktree", "add", "-b", localBranchName, "--", worktreePath, remoteBranch],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> branchExistsLocally({
+    required String workingDirectory,
+    required String branchName,
+  }) {
+    return _processRunner.run(
+      "git",
+      ["branch", "--list", "--", branchName],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> revParse({
+    required String workingDirectory,
+    required String ref,
+  }) {
+    return _processRunner.run(
+      "git",
+      ["rev-parse", ref],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> isAncestor({
+    required String workingDirectory,
+    required String ancestorRef,
+    required String descendantRef,
+  }) {
+    return _processRunner.run(
+      "git",
+      ["merge-base", "--is-ancestor", ancestorRef, descendantRef],
+      workingDirectory: workingDirectory,
+    );
+  }
+
+  Future<ProcessResult> listRemotes({required String workingDirectory}) {
+    return _processRunner.run(
+      "git",
+      ["remote"],
+      workingDirectory: workingDirectory,
+    );
+  }
+
   String? _extractBranchName({required Object? output, required String prefix}) {
     final trimmedOutput = output.toString().trim();
     if (!trimmedOutput.startsWith(prefix)) {
