+++ shared/sesori_shared/lib/sesori_shared.dart
@@ -18,6 +18,8 @@ export "src/models/sesori/active_session.dart";
 export "src/models/sesori/agent_info.dart";
 export "src/models/sesori/agent_mode.dart";
 export "src/models/sesori/base_branch_response.dart";
+export "src/models/sesori/branch_info.dart";
+export "src/models/sesori/branch_list_response.dart";
 export "src/models/sesori/create_session_request.dart";
 export "src/models/sesori/delete_session_request.dart";
 export "src/models/sesori/device_platform.dart";
@@ -51,6 +53,7 @@ export "src/models/sesori/session_status.dart";
 export "src/models/sesori/set_base_branch_request.dart";
 export "src/models/sesori/success_empty_response.dart";
 export "src/models/sesori/update_session_archive_request.dart";
+export "src/models/sesori/worktree_mode.dart";
 export "src/protocol/close_codes.dart";
 export "src/protocol/constants.dart";
 export "src/protocol/framing.dart";
